+++ packages/web3-eth-accounts/src/account.ts
@@ -52,6 +52,7 @@ import {
 	sha3Raw,
 	toChecksumAddress,
 	utf8ToHex,
+	uuidV4,
 } from 'web3-utils';
 import { isBuffer, isNullish, isString, validator } from 'web3-validator';
 import { keyStoreSchema } from './schemas';
@@ -351,37 +352,6 @@ export const recover = (
 	const address = toChecksumAddress(`0x${publicHash.slice(-40)}`);
 
 	return address;
-};
-
-/**
- * Generate a version 4 (random) uuid
- * https://github.com/uuidjs/uuid/blob/main/src/v4.js#L5
- */
-
-const uuidV4 = (): string => {
-	const bytes = randomBytes(16);
-
-	// https://github.com/ethers-io/ethers.js/blob/ce8f1e4015c0f27bf178238770b1325136e3351a/packages/json-wallets/src.ts/utils.ts#L54
-	// Section: 4.1.3:
-	// - time_hi_and_version[12:16] = 0b0100
-	/* eslint-disable-next-line */
-	bytes[6] = (bytes[6] & 0x0f) | 0x40;
-
-	// Section 4.4
-	// - clock_seq_hi_and_reserved[6] = 0b0
-	// - clock_seq_hi_and_reserved[7] = 0b1
-	/* eslint-disable-next-line */
-	bytes[8] = (bytes[8] & 0x3f) | 0x80;
-
-	const hexString = bytesToHex(bytes);
-
-	return [
-		hexString.substring(2, 10),
-		hexString.substring(10, 14),
-		hexString.substring(14, 18),
-		hexString.substring(18, 22),
-		hexString.substring(22, 34),
-	].join('-');
 };
 
 /**